requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bottleneck==1.4.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.1
debugpy==1.8.11
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docstring-parser==0.16
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
google-api-core==2.24.2
google-auth==2.40.0
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain-core==0.3.58
langchain-google-vertexai==2.0.21
langchain-openai==0.3.16
langchain-together==0.3.0
langsmith==0.3.42
latex2sympy2-extended==1.10.1
math-verify==0.7.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numexpr==2.10.2
numpy==2.2.5
openai==1.77.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.1
platformdirs==4.3.7
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0rc2
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==78.1.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
validators==0.35.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
